+++ tests/test_tushare.py
@@ -23,7 +23,8 @@ def main():
     """主函数"""
     get_config()  # 确保配置已加载
     client = TushareClient()
-    df = client.get_realtime_quote("600036.SH")
+    #df = client.get_realtime_quote("600036.SH")
+    df = client.pro.query('forecast_vip', ann_date='20251231')
     print(df)
 
 if __name__ == "__main__":